--- src/components/Button.js
@@ -1,26 +1,0 @@
-import Button from '@material-ui/core/Button';
-import { makeStyles } from '@material-ui/core/styles';
-import PlayArrowIcon from '@material-ui/icons/PlayArrow';
-
-const useStyles = makeStyles((theme) => ({
-  button: {
-    margin: theme.spacing(1),
-  },
-}));
-
-export default function IconLabelButtons() {
-  const classes = useStyles();
-
-  return (
-    <div>
-      <Button
-        variant="contained"
-        color="primary"
-        className={classes.button}
-        endIcon={<PlayArrowIcon />}
-      >
-        Start Long Calculation
-      </Button>
-    </div>
-  );
-}
