+++ src/components/MenuBar/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Bar from './menu-bar';
+
+const MenuBar = () => {
+    return (
+        <Bar title='Machine Learning web app 💡'/>
+    );
+}
+
+export default MenuBar;
